test: move at-timestamp resolution test to history_manager integration
