+++ kernel/tests/integration/log/history_manager.rs
@@ -1,7 +1,10 @@
 //! Integration tests for the public `history_manager` API.
 
+use std::fs::OpenOptions;
 use std::ops::RangeInclusive;
+use std::time::{Duration, SystemTime};
 
+use delta_kernel::committer::FileSystemCommitter;
 use delta_kernel::history_manager::{get_earliest_commit, HistoryCommitType};
 use delta_kernel::object_store::path::Path;
 use delta_kernel::object_store::ObjectStore;
@@ -10,11 +13,15 @@ use delta_kernel::object_store::ObjectStore;
 // silence the resulting unused-import warning.
 #[allow(unused_imports)]
 use delta_kernel::object_store::ObjectStoreExt as _;
-use delta_kernel::Version;
+use delta_kernel::transaction::create_table::create_table;
+use delta_kernel::{DeltaResult, Snapshot, Version};
 use rstest::rstest;
 use test_utils::delta_kernel_default_engine::DefaultEngineBuilder;
-use test_utils::table_builder::{LogState, TestTableBuilder};
+use test_utils::table_builder::{LogState, TestTableBuilder, VersionTarget};
+use test_utils::{build_snapshot, test_table_setup};
 use url::Url;
+
+use crate::common::write_utils::get_simple_int_schema;
 
 async fn remove_commits(store: &dyn ObjectStore, versions: RangeInclusive<Version>) {
     for version in versions {
@@ -55,5 +62,53 @@ async fn test_get_earliest_commit(
         commit_type,
     )?;
     assert_eq!(earliest, expected_version);
+    Ok(())
+}
+
+/// Verifies that [`VersionTarget::AtTimestamp`] resolves through `build_snapshot!` to the
+/// version `latest_version_as_of` selects for a timestamp. `InMemory` collapses successive
+/// `put` timestamps into a single millisecond, so this writes on the local filesystem and
+/// sets each commit's modification time explicitly to get distinct, monotonic commit
+/// timestamps. Resolution correctness across timestamp/commit-layout combinations is covered
+/// by the `history_manager` unit tests; this is the end-to-end `TestTableBuilder` check.
+#[test]
+fn test_at_timestamp_resolves_to_intermediate_version() -> DeltaResult<()> {
+    let (_temp_dir, table_path, engine) = test_table_setup()?;
+
+    // v0: CreateTable
+    let schema = get_simple_int_schema();
+    let mut snap = create_table(&table_path, schema, "AtTimestampTest/1.0")
+        .build(engine.as_ref(), Box::new(FileSystemCommitter::new()))?
+        .commit(engine.as_ref())?
+        .unwrap_post_commit_snapshot();
+
+    // v1..=4: noop commits (each writes a metaData-free, add-free commit JSON).
+    for _ in 1..=4 {
+        snap = test_utils::begin_transaction(snap.clone(), engine.as_ref())?
+            .with_engine_info("AtTimestampTest")
+            .commit(engine.as_ref())?
+            .unwrap_post_commit_snapshot();
+    }
+
+    // Set each commit's mtime to a distinct, monotonic value (in ms).
+    let table_url = delta_kernel::try_parse_uri(&table_path)?;
+    let log_dir = table_url.to_file_path().unwrap().join("_delta_log");
+    for v in 0..=4u64 {
+        let file_path = log_dir.join(format!("{v:020}.json"));
+        let file = OpenOptions::new().write(true).open(&file_path).unwrap();
+        let time = SystemTime::UNIX_EPOCH + Duration::from_millis((v + 1) * 1000);
+        file.set_modified(time).unwrap();
+    }
+
+    // 2500ms lands strictly between v1 (2000) and v2 (3000), so it resolves to v1.
+    let target = VersionTarget::AtTimestamp(2500);
+    let snap_at_ts = build_snapshot!(target, &table_path, engine.as_ref());
+    assert_eq!(snap_at_ts.version(), 1);
+
+    // i64::MAX resolves to latest.
+    let target_max = VersionTarget::AtTimestamp(i64::MAX);
+    let snap_max = build_snapshot!(target_max, &table_path, engine.as_ref());
+    assert_eq!(snap_max.version(), 4);
+
     Ok(())
 }
